% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASELINE_MODEL.R
\name{BASELINE_MCMC_ADAPTIVE}
\alias{BASELINE_MCMC_ADAPTIVE}
\title{MCMC adaptive algorithm for a baseline epidemic model}
\usage{
BASELINE_MCMC_ADAPTIVE(
  epidemic_data,
  mcmc_inputs = list(n_mcmc = 10000, r0_start = 1.2, r0_prior_exp = c(1, 0),
    target_accept_rate = 0.4, thinning_factor = 10),
  FLAGS_LIST = list(ADAPTIVE = TRUE, PRIOR = TRUE, THIN = TRUE)
)
}
\arguments{
\item{epidemic_data}{data from the epidemic, namely daily infection counts}

\item{mcmc_inputs}{A list of MCMC specifications including
\itemize{
  \item \code{"n_mcmc"} - Number of iterations of the MCMC sampler.
  \item \code{"r0_start"} - Model parameter starting points; where the MCMC algorithm begins sampling the reproduction number R0 from.
  \item \code{"r0_prior_exp"} - Prior on R0. Default prior is an exponential distribution with rate 1.
  \item \code{"target_accept_rate"} - target acceptance rate. Used in the adaptive algorithm
  \item \code{"thinning_factor"}  - factor of total \code{"n_mcmc"} size of which samples are kept. Only if  \code{"FLAGS_LIST$THIN = TRUE"}, otherwise all samples are kept.
}}

\item{FLAGS_LIST}{A list of Boolean variables for switching on/off certain functionality
\itemize{
  \item \code{"ADAPTIVE"} - Adaptive MCMC algorithm implemented if TRUE
  \item \code{"PRIOR"}  - Apply prior distributions to model parameters
  \item \code{"THIN"}  - Return a thinned MCMC sample if TRUE, reduced by a factor of \code{"thinning_factor"}
}}
}
\value{
MCMC samples, acceptance rates etc.
\itemize{
  \item \code{"r0_vec"} - A vector containing MCMC samples of R0, the repoduction number of the epidemic.
  \item \code{"log_like_vec"}  - A vector containing the log likelihood at each of the MCMC iterations, of size of \code{"mcmc_vec_size"}.
  \item \code{"sigma_vec"} - A vector containing sigma adapted at each iteration of the MCMC algorithm, of size of \code{"mcmc_vec_size"} if \code{"FLAGS_LIST$ADAPTIVE"} is TRUE. Otherwise sigma_vec is a constant value.
  \item \code{"accept_rate"}  - The MCMC acceptance rate (percentage) for R0. 
  }
}
\description{
MCMC algorithm with adaptation for obtaining samples of the reproduction number R0 of an epidemic model
}
\examples{

mcmc_inputs = list(n_mcmc = 500000, r0_start = 1.1, r0_prior_exp = c(1, 0), target_accept_rate = 0.4, thinning_factor = 10)

#START MCMC
mcmc_baseline_output = SSE_MCMC_ADAPTIVE(epidemic_data, mcmc_inputs)


}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
