% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_COMPARISON.R
\name{get_bayes_factor_harmonic_means}
\alias{get_bayes_factor_harmonic_means}
\title{Bayes Factor via Harmonic mean}
\usage{
get_bayes_factor_harmonic_means(loglike_vec1, loglike_vec2)
}
\arguments{
\item{log_like_vec1}{Log Likelihood of model 1 at each iteration of the MCMC algorithm}

\item{log_like_vec2}{Log Likelihood of model 2 at each iteration of the MCMC algorithm}
}
\value{
\code{"bayes_factor"} bayes factor of two models up for comparison. Obtained via their harmonic means
}
\description{
Bayes Factor of two competing models obtained via the ratio of their two harmonic means
}
\examples{

bayes_factor = get_bayes_factor_hm(loglike_vec1, loglike_vec2)


}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
