% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_CRITICISM.R
\name{GET_SUM_STATS_TOTAL}
\alias{GET_SUM_STATS_TOTAL}
\title{Get and save the summary statistics of the real and replicated data}
\usage{
GET_SUM_STATS_TOTAL(
  epidemic_data,
  root_folder,
  model_type = list(FLAG_BASE = TRUE, FLAG_SSE = FALSE, FLAG_SSI = FALSE),
  modelling_specs = list(n_reps = 10, n_mcmc = 10000, thinning_factor = 10,
    burn_in_size = 0.01, FLAG_THIN = TRUE)
)
}
\arguments{
\item{epidemic_data}{data from the epidemic, namely daily infection counts}

\item{root_folder}{root folder location in which to store the MCMC results for \code{"n_reps"}}

\item{modelling_specs}{A list of specifications;
\itemize{
  \item \code{"n_reps"} - Number of repetitions of the MCMC sampler. Multiple runs required for model criticism
  \item \code{"n_mcmc"} - Number of iterations of the mcmc sampler }
  \item \code{"burn_in_pc"} - Proportion of mcmc samples to remove at the start as burn-in
  \item \code{"thinning_factor"}  - The data is replicated using a thinned sample of the MCMC output}
}
\value{
saves the \code{"df_ss_true_rep_"} and \code{"df_summary_stats"} int the \code{"root_foler"}/\code{"rep"} location
}
\description{
Get and save the summary statistics of the real and replicated data. The data is replicated using a thinned sample of the MCMC output.
 The summary statistics of each replicated dataset are calculated and stored together in a dataframe.
}
\examples{

GET_SUM_STATS_TOTAL(epidemic_data, root_folder)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
