% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_CRITICISM.R
\name{GET_SUMMARY_STATS}
\alias{GET_SUMMARY_STATS}
\title{Calculate a set of summary statistics of the data
Returns the summary statistics of the data in a dataframe or list format}
\usage{
GET_SUMMARY_STATS(epidemic_data, FLAG_DF_CREATE = TRUE)
}
\arguments{
\item{epidemic_data}{Data for which the summary statistics are calculated}

\item{FLAG_DF_CREATE}{Flag to indicate whether to return a dataframe or a list of summary statistics}
}
\value{
summary statistics of the data in a dataframe or list format
}
\description{
Calculate a set of summary statistics of the data
Returns the summary statistics of the data in a dataframe or list format
}
\examples{

summary_stats_results = GET_SUMMARY_STATS(data, FLAG_DF_CREATE = TRUE)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
