% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASELINE_MODEL.R
\name{LOG_LIKE_BASELINE}
\alias{LOG_LIKE_BASELINE}
\title{Log likelihood of the baseline epidemic model}
\usage{
LOG_LIKE_BASELINE(epidemic_data, R0)
}
\arguments{
\item{epidemic_data}{data from the epidemic, namely daily infection counts}

\item{R0}{Reproduction number R0}
}
\value{
Log likelihood of the baseline model
}
\description{
Returns the log likelihood of the baseline epidemic model for given \code{"epidemic_data"} and reproduction number \code{"R0"}
}
\examples{
R0 = 1.2
log_likelihood_baseline = LOG_LIKE_BASELINE(epidemic_data, R0)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
