% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSE_MCMC.R
\name{LOG_LIKE_SSE_LSE}
\alias{LOG_LIKE_SSE_LSE}
\title{Log likelihood of the Super-Spreading Individuals (SSI) epidemic model}
\usage{
LOG_LIKE_SSE_LSE(x, alphaX, betaX, gammaX)
}
\arguments{
\item{epidemic_data}{data from the epidemic, namely daily infection counts}

\item{alpha_X}{SSI model parameter \code{"\alpha"}}

\item{beta_X}{SSI model parameter \code{"\beta"}}

\item{gamma_X}{SSI model parameter \code{"\gamma"}}
}
\value{
Log likelihood of the SSE model
}
\description{
Returns the Log likelihood of the super-spreading individuals (SSI)epidemic model for given \code{"epidemic_data"} and set of model parameters \code{"a, b"} and \code{"c"} and
}
\examples{

log_like = LOG_LIKE_SSE_LSE(epidemic_data, 0.8, 0.02, 20)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
