% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ss_grid.R
\name{PLOT_SS_MCMC_GRID}
\alias{PLOT_SS_MCMC_GRID}
\title{Grid Plot of MCMC Super-Spreading Results}
\usage{
PLOT_SS_MCMC_GRID(
  epidemic_data,
  mcmc_output,
  mcmc_specs = list(n_mcmc = n_mcmc, burn_in_size = 0.05, mod_start_points =
    mod_start_points, mod_par_names = c("a", "b", "c"), model_type = "SSI", seed_count =
    seed_count, thinning_factor = 10),
  priors_list = list(a_prior_exp = c(1, 0), b_prior_ga = c(10, 2/100), b_prior_exp =
    c(0.1, 0), c_prior_ga = c(10, 1), c_prior_exp = c(0.1, 0)),
  FLAGS_LIST = list(BURN_IN = TRUE, THIN = FALSE, DATA_AUG = TRUE, ADAPTIVE = FALSE,
    MULTI_ALG = FALSE, SSI = FALSE, PRIOR = TRUE, B_PRIOR_GAMMA = TRUE, C_PRIOR_GAMMA =
    TRUE)
)
}
\arguments{
\item{epidemic_data}{data from the epidemic, namely daily infection counts}

\item{mcmc_output}{mcmc samples from mcmc sampler/algorithm}

\item{mcmc_specs}{A list of mcmc specifications
\itemize{
  \item \code{"n_mcmc"} - Number of iterations of the mcmc sampler (integer)
  \item \code{"burn_in_size"} - Proportion of mcmc samples to remove at the start as burn-in
  \item \code{"mod_start_points"} - Model parameter starting points; where the mcmc algorithm started sampling from
  \item \code{"mod_par_names"} - Names of the model parameters, e.g \code{"a, b, c"}
  \item \code{"model_type"} - Model type; Super Spreading Individuals or Events
  \item \code{"seed_count"} - Seed for data generation & mcmc iteration
  \item \code{"thinning_factor"}  - factor of total \code{"n_mcmc"} size of which samples are kept. Only if  \code{"FLAGS_LIST$THIN = TRUE"}, otherwise all samples are kept
}}

\item{priors_list}{A list of prior parameters used
\itemize{
  \item \code{"a_prior_exp"} - rate of exponential prior on a, default rate of 1
  \item \code{"b_prior_exp"} - rate of exponential prior on b, default rate of 1
  \item \code{"b_prior_ga"} -  shape and scale of gamma distriubtion prior on b, defaults Ga(10, 0.02)
  \item \code{"c_prior_exp"}  - rate of exponential prior on c, default rate of 0.1
  \item \code{"c_prior_ga"} -  shape and scale of gamma distriubtion prior on c, defaults Ga(10, 1)
}}

\item{FLAGS_LIST}{A list of boolean variables for switching on/off certain functionality
\itemize{
\item \code{"BURN_IN"}  - Burn-in applied to mcmc samples if TRUE of size \code{"mcmc_specs$burn_in_size"}
  \item \code{"THIN"}  - Return a thinned mcmc sample if TRUE, reduced by a factor of \code{"thinning_factor"}
  \item \code{"DATA_AUG"} - Data Augmentation was implemented as part of the SSI model if TRUE
  \item \code{"ADAPTIVE"} - Adaptive Algorithm applied to MCMC samples if TRUE
  \item \code{"MULTI_ALG"}  -  Multi Adaptive Shaping Algorithm applied to MCMC samples if TRUE
  \item \code{"PRIOR"}  - Plot prior distributions in grid if TRUE
  \item \code{"B_PRIOR_GAMMA"}  - A Gamma prior on b if TRUE, otherwise exponential
  \item \code{"C_PRIOR_GAMMA"}  - A Gamma prior on c if TRUE, otherwise exponential
}}
}
\value{
Dataframe of results including mcmc sample starting points, mcmc sample final means, acceptance rates, mcmc effective sizes and the mcmc sampler run time
}
\description{
Grid Plot of MCMC Results for Super-Spreading Epidemic Models
}
\examples{

mcmc_plot_inputs = list(n_mcmc = 500000,  burn_in_pc = 0.05, mod_start_points = list(m1 = 0.72, m2 = 0.0038, m3 = 22),
mod_par_names = c('a', 'b', 'c'), model_type = 'SSI', seed_count = seed_count, thinning_factor = 10)

df_mcmc_results = PLOT_SS_MCMC_GRID(epidemic_data, mcmc_output)
}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
