% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_CRITICISM.R
\name{RUN_MODEL_CRITICISM}
\alias{RUN_MODEL_CRITICISM}
\title{Perform Model Criticism of an epidemic model via posterior predictive checking}
\usage{
RUN_MODEL_CRITICISM(
  epidemic_data,
  root_folder,
  model_type = list(FLAG_BASE = TRUE, FLAG_SSE = FALSE, FLAG_SSI = FALSE),
  modelling_specs = list(n_reps = 10, n_mcmc = 1000)
)
}
\arguments{
\item{epidemic_data}{data from the epidemic, namely daily infection counts}

\item{root_folder}{root folder location in which to store results}

\item{modelling_specs}{A list of specifications;
\itemize{
  \item \code{"n_reps"} - Number of repetitions of the mcmc sampler. Multiple runs required for model criticism
  \item \code{"n_mcmc"} - Number of iterations of the mcmc sampler }}
}
\value{
saves the \code{"mcmc_output"} int the \code{"root_foler"}/\code{"rep"} location
}
\description{
Model Criticism of an epidemic model via posterior predictive checking. Calls a series of four functions to execute the process fully.
}
\examples{

RUN_MODEL_CRITICISM(epidemic_data, root_folder, model_type = list(FLAG_BASE = FALSE, FLAG_SSE = FALSE, FLAG_SSI = TRUE))


}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
