% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMULATE_EPIDEMICS.R
\name{SIMULATE_SS_EVENTS}
\alias{SIMULATE_SS_EVENTS}
\title{Simulate an epidemic with super-spreading events \code{'SSE'}}
\usage{
SIMULATE_SS_EVENTS(
  alphaX,
  betaX,
  gammaX,
  num_days = 50,
  shape_gamma = 6,
  scale_gamma = 1
)
}
\arguments{
\item{num_days}{Number of days of the epidemic}

\item{shape_gamma}{Shape of the gamma distribution representing the time-varying infectivity curve of each infected individual}

\item{scale_gamma}{Scale of the gamma distribution representing the time-varying infectivity curve of each infected individual}

\item{alpha_X}{Model parameter \code{"alpha"}. The daily infection count from regular spreading is assumed to follow a poisson distribution with rate \code{alpha}*\code{\lambda_t}}

\item{beta_X}{SSE model parameter \code{"beta"}.}

\item{gamma_X}{SSE model parameter \code{"gamma"}}
}
\value{
Total infections; Total daily infection counts of length \code{'num_days'}
}
\description{
Returns daily infection counts from a simulated epidemic outbreak
}
\examples{

tot_daily_sse_infections = SIMULATE_SS_EVENTS(num_days = 50, shape_gamma = 6, scale_gamma = 1,
 alphaX = 0.8, betaX = 0.02, gammaX = 20)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
