% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSI_MULTI_MCMC.R
\name{SSI_MULTI_MCMC_ADADPTIVE}
\alias{SSI_MULTI_MCMC_ADADPTIVE}
\title{MCMC algorithm for Super-Spreading Individuals (SSI) epidemic model}
\usage{
SSI_MULTI_MCMC_ADADPTIVE(
  data,
  mcmc_inputs = list(n_mcmc = n_mcmc, mod_start_points = mod_start_points, burn_in_pc =
    0.05, dim = 3, alpha_star = 0.4, v0 = 100, vec_min = c(0, 0, 0), thinning_factor =
    10),
  priors_list = list(b_prior_exp = c(1, 0), c_prior_exp = c(0.1, 0)),
  FLAGS_LIST = list(DATA_AUG = TRUE, B_PRIOR_EXP = TRUE, C_PRIOR_EXP = TRUE, THIN =
    TRUE)
)
}
\arguments{
\item{data}{data from the epidemic, namely daily infection counts}

\item{mcmc_inputs}{A list of mcmc specifications including
\itemize{
  \item \code{"n_mcmc"} - Number of iterations of the mcmc sampler
  \item \code{"mod_start_points"} - Model parameter starting points; where the mcmc algorithm begins sampling from
  \item \code{"burn_in_pc"} - Burn-in percentage; the percentage of the total mcmc iterations disgarded as burn-in
  \item \code{"dim"} - dimension of model or number of unknown model parameters
  \item \code{"alpha_star"} - target acceptance rate; used in the adaptive algorithm
  \item \code{"v0"} - used in the adaptive algorithm
  \item \code{"vec_min"} - vector specifying the minimum accepted size of each of the model parameters
  \item \code{"thinning_factor"}  - factor of total \code{"n_mcmc"} size of which samples are kept. Only if  \code{"FLAGS_LIST$THIN = TRUE"}, otherwise all samples are kept
}}

\item{priors_list}{A list of prior parameters used
\itemize{
  \item \code{"b_prior_exp"} - rate of exponential prior on b, default rate of 1
  \item \code{"c_prior_exp"}  - rate of exponential prior on c, default rate of 0.1
}}

\item{FLAGS_LIST}{A list of boolean variables for switching on/off certain functionality
\itemize{
  \item \code{"DATA_AUG"} - Data Augmentation implemented as part of the SSI model if TRUE
  \item \code{"B_PRIOR_EXP"}  - An exponential prior on b if TRUE
  \item \code{"C_PRIOR_EXP"}  - An exponential prior on c if TRUE
  \item \code{"THIN"}  - Return a thinned mcmc sample if TRUE, reduced by a factor of \code{"thinning_factor"}
}}
}
\value{
MCMC samples & other output
\itemize{
  \item \code{"x_matrix"} - A matrix of size \code{"mcmc_vec_size"} x \code{"dim"}, i.e each column corresponds to the mcmc samples of a model parameter
  \item \code{"log_like_vec"}  - A vector containing the log likelihood at each of the mcmc iterations, of size of \code{"mcmc_vec_size"}
  \item \code{"lambda_vec"}  - A vector containing the value of lambda in the adaptive algorithm at each of the mcmc iterations, of \code{"mcmc_vec_size"}
  \item \code{"accept_rate"}  - The acceptance rate of the mcmc algorithm
  \item \code{"accept_rate_da"}  - The acceptance rate of the data augmentation step in the mcmc algorithm
  \item \code{"non_ss"}  - The augmented epidemic data corresponding to the non super-spreaders (non-ss) for each iteration of the mcmc
  \item \code{"ss"}  - The augmented epidemic data corresponding to the super-spreaders (ss) for each iteration of the mcmc
  }
}
\description{
MCMC algorithm for obtaining samples from the parameters of a
super-spreading individuals epidemic model
}
\examples{

mcmc_inputs = list(n_mcmc = 500000, mod_start_points = list(m1 = 0.72, m2 = 0.0038, m3 = 22), burn_in_pc = 0.05,
dim = 3, alpha_star = 0.4, v0 = 100, vec_min = c(0,0,1), thinning_factor = 10)

#START MCMC
mcmc_ssi_output = MCMC_SSI_MULTI_ADADPT(epidemic_data, mcmc_inputs)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
